+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-09-07 02:56:11
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 17150372dead26df709fc761b2c643aa7a3d6d05
+
+## [ForwardWidgets-2kuai] - 2025-09-07 02:56:11
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: e851dfd71d46be56e5f34456a07f52d23bb808dd
+
+## [ForwardWidgets-pack1r] - 2025-09-07 02:56:11
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 0a7250578902c5d1da8887f4fb0ca9c75837401d
+
+
 ## [ForwardWidgets-huangxd] - 2025-09-06 02:44:58
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main